Creating a semantic group
This is a code to generates a semantic group

The aim is to start from one word and to find the word most correlated with that one. Then to download the word mostly correlated with the second one, then the words mostly correlated with the third one, etc, etc. Until no new words are found.

For example:
0: set of words = ['Furniture]
1: Download the tweets that on a given day containing "#furniture"
2: Find the most highly ranked hashtags in this set. For example: "#design" and "#architecture".
3: Add the term "design" in the initial set: ['Furniture', 'Design']
4: Go back to step 1 using the hashtag "#Design'

Repeat until no new words are added to the set.